requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning extensions
imbalanced-learn>=0.9.0
lightgbm>=3.3.0
xgboost>=1.6.0
catboost>=1.0.0

# Survival analysis
scikit-survival>=0.19.0

# Deep learning
tensorflow>=2.8.0
torch>=1.11.0

# Data visualization
plotly>=5.6.0
dash>=2.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
pyyaml>=6.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0
